split algolia deploy steps
